+++ internal/db/model/Organization.go
@@ -9,8 +9,8 @@ import (
 
 // Organization represents an organization account.
 type Organization struct {
-	ID           int64     `db:"id"`
-	ExternalID   int64     `db:"extern_id"`
+	ID           uint64    `db:"id"`
+	ExternalID   uint64    `db:"extern_id"`
 	Provider     string    `db:"provider"`
 	Login        string    `db:"login"`
 	AvatarURL    string    `db:"avatar_url"`
@@ -28,8 +28,8 @@ type Organization struct {
 // ToDto converts model to v1.Organization DTO
 func (u *Organization) ToDto() *v1.Organization {
 	user := &v1.Organization{
-		ID:           strconv.FormatUint(uint64(u.ID), 10),
-		ExternalID:   strconv.FormatUint(uint64(u.ExternalID), 10),
+		ID:           strconv.FormatUint(u.ID, 10),
+		ExternalID:   strconv.FormatUint(u.ExternalID, 10),
 		Provider:     u.Provider,
 		Login:        u.Login,
 		Name:         u.Name,
